FROMLIST: arm64: dts: msm8916-samsung-a2015: add tactile buttons and hall sensor
Add nodes for basic GPIO connected hardware to the Samsung A3/A5 common dtsi.
This includes the Volume UP button, the Home button, and the hall sensor used
to sense "smart cover" open state. Related to that, add a node for the Volume
DOWN button, which is handled by the pm8916 as is common with msm8916 devices.

Signed-off-by: Michael Srba <Michael.Srba@seznam.cz>