FROMLIST: ASoC: msm8916-wcd-digital: Remove broken MIX2 DAPM routes
Since 5.3-rc1, the kernel (correctly) complains about missing
widgets when loading msm8916-wcd-digital:

	ASoC: no sink widget found for RX1 MIX2 INP1
	ASoC: Failed to add route IIR1 -> IIR1 -> RX1 MIX2 INP1
	ASoC: no sink widget found for RX2 MIX2 INP1
	ASoC: Failed to add route IIR1 -> IIR1 -> RX2 MIX2 INP1
	ASoC: no sink widget found for RX1 MIX2 INP1
	ASoC: Failed to add route IIR2 -> IIR2 -> RX1 MIX2 INP1
	ASoC: no sink widget found for RX2 MIX2 INP1
	ASoC: Failed to add route IIR2 -> IIR2 -> RX2 MIX2 INP1

This happens because RX1/2 MIX2 INP1 is not actually supported by
msm8916-wcd-digital for some reason. There were some traces of it
in the initial patch, but the last remnants were removed in
commit fecd5c09ddf8 ("ASoC: msm8916-wcd-digital: remove few unused variables").

On 5.3 everything seemingly worked even with that error, but on 5.4-rc1
the sound card is no longer successfully registered:

	qcom-apq8016-sbc 7702000.sound: ASoC: failed to instantiate card -19

This happens because a check for the return code of
snd_soc_dapm_add_routes() was added in
commit daa480bde6b3 ("ASoC: soc-core: tidyup for snd_soc_dapm_add_routes()").

Avoid the error by removing the routes that reference the missing widgets.

Fixes: ef8a4757a6db ("ASoC: msm8916-wcd-digital: Add sidetone support")
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>