clk: qcom: apcs-msm8916: add 'xo' clock into parent map
APCS clock of MSM8939 cluster0 is initially running from 'xo' source.
With it being absent from parent map will cause failure in
clk-regmap-mux-div operations like below.

 [    0.893176] a53mux_c0: Can't find parent with src 0
 [    0.893243] a53mux_c0: Can't find parent 0

Let's add it into parent map to fix the failure.  While at it, rename
the parent map and name variable to be generic, so that we do not have
to change every time we add or remove element from the parent map.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>